% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.pf.R
\name{mean.pf}
\alias{mean.pf}
\title{Mean method for objects of class pf}
\usage{
\method{mean}{pf}(x, bylocus = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{pf}: output of \link{pr_total_number_of_distinct_alleles}}

\item{bylocus}{If \code{TRUE} then the results will be returned locus by locus}

\item{...}{other arguments that may p}
}
\value{
either a vector of means, one for each locus, or the sum of the locus means.
The means are the expected number of alleles observed at each locus
}
\description{
Mean method for objects of class pf
}
\examples{
data(FBIfreqs)
p <- pr_total_number_of_distinct_alleles(contributors = c("U1","U2"),
                                         freqs = FBIfreqs)
mean(p)
var(p)
}
