% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberofalleles.R
\docType{package}
\name{numberofalleles}
\alias{numberofalleles}
\title{numberofalleles}
\description{
Probabilities for observing a number of distinct alleles in a forensic DNA mixture
}
\author{
Maarten Kruijver
}
