% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pf.R
\name{plot.pf}
\alias{plot.pf}
\title{Plot method for objects of class pf}
\usage{
\method{plot}{pf}(x, plot.lines = TRUE, plot.points = TRUE, ...)
}
\arguments{
\item{x}{An object of class pf}

\item{...}{Any other arguments that should be sent to \codeP}
}
\description{
Plot method for objects of class pf
}
