% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_number_of_distinct_alleles.R
\name{pr_number_of_distinct_alleles}
\alias{pr_number_of_distinct_alleles}
\title{Compute probability distribution of number of distinct alleles in a DNA mixture at a locus}
\usage{
pr_number_of_distinct_alleles(number_of_alleles, f)
}
\arguments{
\item{number_of_alleles}{Integer. Number of (not necessarily distinct) alleles in the mixture.}

\item{f}{Numeric Vector with allele frequencies}
}
\description{
Compute probability distribution of number of distinct alleles in a DNA mixture at a locus
}
\details{
TODO
}
\examples{
f <- c(A = 0.1, B = 0.2, C = 0.7)

pr_number_of_distinct_alleles(3, f)
}
