% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_total_number_of_alleles.R
\name{pr_total_number_of_distinct_alleles}
\alias{pr_total_number_of_distinct_alleles}
\title{Compute probability distribution of total number of distinct alleles in a DNA mixture}
\usage{
pr_total_number_of_distinct_alleles(
  contributors,
  freqs,
  pedigree,
  loci = names(freqs)
)
}
\arguments{
\item{contributors}{Character vector with unique names of contributors. Valid names are "U1", "U2", ... for unrelated contributors or the names of pedigree members for related contributors.}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{pedigree}{(optionally) \link[pedtools]{ped} object}

\item{loci}{Character vector of locus names (defaults to names attr. of \code{freqs})}
}
\description{
Compute probability distribution of total number of distinct alleles in a DNA mixture
}
\details{
TODO
}
\examples{
# define a pedigree of siblings S1 and S2 (and their parents)
ped_sibs <- pedtools::nuclearPed(nch = 2,
father = "F", mother = "M",
children = c("S1", "S2"))

# define allele frequencies
freqs <- list(locus1 = c(0.1, 0.9),
              locus2 = c(0.25, 0.5, 0.25))

# compute dist. of number of alleles for two siblings and one unrelated persons
pr_total_number_of_distinct_alleles(contributors = c("S1","S2","U1"), freqs, ped_sibs)

## GlobalFiler example
freqs <- read_allele_freqs(system.file("extdata","FBI_extended_Cauc.csv",
package = "numberofalleles"))

gf_loci <- c("D3S1358", "vWA", "D16S539", "CSF1PO", "TPOX", "D8S1179",
             "D21S11",  "D18S51", "D2S441", "D19S433", "TH01", "FGA",
             "D22S1045", "D5S818", "D13S317", "D7S820", "SE33",
             "D10S1248", "D1S1656", "D12S391", "D2S1338")

p_gf <- pr_total_number_of_distinct_alleles(contributors = c("U1", "U2"),
                                            freqs = freqs, loci = gf_loci)

barplot(p_gf)
}
