% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pr_number_of_distinct_alleles.R
\name{estimate_pr_number_of_distinct_alleles}
\alias{estimate_pr_number_of_distinct_alleles}
\title{Estimate the probability distribution of the number of distinct alleles observed in a DNA mixture for a locus}
\usage{
estimate_pr_number_of_distinct_alleles(
  number_of_independent_alleles,
  f,
  fst = 0,
  number_of_samples = 1e+05,
  importance_sampling = FALSE
)
}
\arguments{
\item{number_of_independent_alleles}{Integer. Number of independent alleles in the mixture.}

\item{f}{Numeric vector with allele frequencies}

\item{fst}{Numeric value for sub-population correction (also known as theta)}
}
\value{
A named numeric vector describing the probability distribution. Numeric values are the probabilities corresponding to the names describing integer values.
}
\description{
For a given number of \emph{independent} alleles, estimates the probability distribution of the number of \emph{distinct} alleles observed in a DNA mixture.
}
\details{
Due to allele sharing between DNA mixture contributors, the number of \emph{distinct} alleles observed in a mixture is often less than the number of independent alleles in the mixture. For example, if mixture comprises two unrelated contributors, there are four independent alleles. Some of these four independent alleles may be of the same allelic type so that at least one and at most four distinct alleles are observed.

This function estimates the probability distribution of the number of \emph{distinct} alleles observed when the mixtures comprises a given number of \emph{independent} alleles. Optionally, a sub-population correction may be applied by setting \code{fst>0}.

The function uses a Monte Carlo estimation procedure.
}
\examples{

freqs <- read_allele_freqs(system.file("extdata",
                                      "FBI_extended_Cauc.csv",
                                      package = "numberofalleles"))

# compute the exact distribution of distinct alleles for a 5-person mixture at SE33
p <- pr_number_of_distinct_alleles(number_of_independent_alleles = 10,
                                  f = freqs$SE33)

plot(seq(p), p, type = "b", log = "y", xlab = "Number of distinct alleles")
grid()

# next we compare it to a simple Monte Carlo estimate and the Importance Sampling estimate
set.seed(1)
p_hat <- estimate_pr_number_of_distinct_alleles(number_of_independent_alleles = 10,
                                               f = freqs$SE33, number_of_samples = 1e4)
p_hat_is <- estimate_pr_number_of_distinct_alleles(number_of_independent_alleles = 10,
                                                  f = freqs$SE33, number_of_samples = 1e4,
                                                  importance_sampling = TRUE)

# Monte Carlo fails to estimate the tiny probabilities
# but Importance Sampling is fairly accurate even with only 10k samples
plot(seq(p), p_hat/p, type="b", lty=2, xlab="Number of distinct alleles")
lines(p_hat_is/p, type="b")
grid()
legend("bottomright", legend=c("Monte Carlo", "Importance Sampling"), lty = c(2,1))

}
