% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.pf.R
\name{mean.pf}
\alias{mean.pf}
\title{Mean method for objects of class pf}
\usage{
\method{mean}{pf}(x, bylocus = FALSE, ...)
}
\arguments{
\item{bylocus}{If \code{TRUE} then the results will be returned locus by locus}

\item{...}{other arguments that may p}
}
\value{
either a vector of means, one for each locus, or the sum of the locus means.
The means are the expected number of alleles observed at each locus
}
\description{
Mean method for objects of class pf
}
