% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_independent_alleles_ped.R
\name{pr_independent_alleles_ped}
\alias{pr_independent_alleles_ped}
\title{Compute the probability distribution of the number of independent alleles in a mixture with dropout and related contributors}
\usage{
pr_independent_alleles_ped(pedigree, ped_contributors, dropout_prs)
}
\arguments{
\item{pedigree}{\link[pedtools]{ped} object}

\item{ped_contributors}{Character vector with unique names of contributors. Valid names are the names of pedigree members.}

\item{dropout_prs}{Numeric vector. Dropout probabilities per contributor.}
}
\description{
When mixture contributors are related according to a pedigree, they may share some alleles identical by descent so that their total number of \emph{independent} alleles is smaller than two times the number of contributors. The number of \emph{independent} alleles can be further reduced if dropout plays a role. This function computes the probability distribution of the number of independent alleles that related mixture contributors have in total for a locus given their dropout parameters. Note that the number of \emph{distinct} alleles that is observed at the locus is typically smaller than the number of independent alleles due to allele sharing.
}
\examples{
# a dropout pr. of 0.5
p <- pr_independent_alleles(0.5)
stopifnot(all.equal(as.vector(p),
                    c(0.5^2, 2 * 0.5 * 0.5, (1-0.5)^2)))

# one contrib. without dropout and one with d=0.5
p1 <- pr_independent_alleles(c(0, 0.5))
stopifnot(identical(as.integer(names(p1)),
                    as.integer(names(p)) + 2L))
}
\seealso{
\link{pr_independent_alleles}
}
