% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_total_number_of_alleles.R
\name{pr_total_number_of_distinct_alleles}
\alias{pr_total_number_of_distinct_alleles}
\title{Compute probability distribution of total number of distinct alleles in a DNA mixture}
\usage{
pr_total_number_of_distinct_alleles(
  contributors,
  freqs,
  pedigree,
  loci = names(freqs)
)
}
\arguments{
\item{contributors}{Character vector with unique names of contributors. Valid names are "U1", "U2", ... for unrelated contributors or the names of pedigree members for related contributors.}

\item{freqs}{Allele frequencies (see \link{read_allele_freqs})}

\item{pedigree}{(optionally) \link[pedtools]{ped} object}

\item{loci}{Character vector of locus names (defaults to names attr. of \code{freqs})}
}
\description{
Compute probability distribution of total number of distinct alleles in a DNA mixture
}
\details{
TODO
}
\examples{
# define a pedigree of siblings S1 and S2 (and their parents)
ped_sibs <- pedtools::nuclearPed(nch = 2,
father = "F", mother = "M",
children = c("S1", "S2"))

# define allele frequencies
freqs <- list(locus1 = c(0.1, 0.9),
              locus2 = c(0.25, 0.5, 0.25))

pr_total_number_of_distinct_alleles(contributors = c("S1","S2","U1"), freqs, ped_sibs)
}
